requirements.txt
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
colorama==0.4.4
cryptography==36.0.1
Deprecated==1.2.13
idna==3.3
loguru==0.5.3
pyaes==1.6.1
pyasn1==0.4.8
pycparser==2.21
PyJWT==2.3.0
pyTelegramBotAPI==4.3.1
python-decouple==3.5
python-telegram-handler==2.2
pytz==2021.3
requests==2.27.1
rsa==4.8
Telethon==1.24.0
urllib3==1.26.8
vonage==2.5.5
win32-setctime==1.0.4
wrapt==1.13.3
